feat : Outputview를 좀 더 상세화
